Wip: nn
it works but ive yet to add other hyperparameters.
so far, the only hyperparameters i've considered are leaky relu vs relu and the number of hidden layers.

note that i cut the epoch to 50 just to test if the loop is working. you may want to turn this up to 300 (as seen in the original notebook)